pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "atlas-video"
version = "0.1.0"
description = "A multimodal insights engine for video understanding"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.12"
authors = [
    { name = "VeedoAI", email = "hello@veedoai.com" }
]
keywords = [
    "video",
    "multimodal",
    "ai",
    "gemini",
    "transcription",
    "video-analysis",
    "vector-search",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Video",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Text Processing :: Linguistic",
]
dependencies = [
    "click>=8.0.0",
    "pydantic>=2.0.0",
    "google-genai>=1.0.0",
    "groq>=0.4.0",
    "zvec>=0.2.0",
    "rich>=13.0.0",
    "httpx>=0.25.0",
    "numpy>=1.20.0",
    "python-dotenv",
    "nanoid"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
atlas = "atlas.cli:main"

[project.urls]
Homepage = "https://github.com/veedoai/atlas"
Documentation = "https://github.com/veedoai/atlas#readme"
Repository = "https://github.com/veedoai/atlas"
Issues = "https://github.com/veedoai/atlas/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/atlas"]

[tool.hatch.build.targets.sdist]
include = ["src/atlas", "README.md", "LICENSE"]

[tool.ruff]
line-length = 120
target-version = "py312"
include = ["*.py", "*.pyi", ".pyproject.toml"]

[tool.ruff.lint]
select = ["D102", "F401"]

[tool.ruff.lint.isort]
case-sensitive = true
relative-imports-order = "closest-to-furthest"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pylint]
max-line-length = 120

[tool.mypy]
ignore_missing_imports = false
python_version = "3.12"
strict = true
